Extract Die.Roll to separate file
